requirements.txt
openai==0.28.1
python-dotenv==1.0.0
click==8.1.7 